requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
openai==1.3.0
google-generativeai==0.3.1
Werkzeug==3.0.1
email-validator==2.1.0
gunicorn==21.2.0